simplify GemEIP712Message
