+++ gemstone/src/message/decoder.rs
@@ -2,14 +2,12 @@ use alloy_primitives::hex;
 use bs58;
 
 use crate::GemstoneError;
+use gem_evm::eip712::parse_eip712_json;
 
-use super::{eip712::GemEIP712Message, sign_type::SignDigestType};
-
-#[derive(Debug, uniffi::Record)]
-pub struct SignMessage {
-    pub sign_type: SignDigestType,
-    pub data: Vec<u8>,
-}
+use super::{
+    eip712::GemEIP712Message,
+    sign_type::{SignDigestType, SignMessage},
+};
 
 #[derive(Debug, uniffi::Enum)]
 pub enum MessagePreview {
@@ -52,6 +50,34 @@ impl SignMessageDecoder {
         }
     }
 
+    pub fn hash(&self) -> Vec<u8> {
+        match self.message.sign_type {
+            SignDigestType::Sign => self.message.data.clone(),
+            SignDigestType::Eip191 => {
+                let prefix = "\x19Ethereum Signed Message:\n";
+                let mut data = prefix.as_bytes().to_vec();
+                data.extend_from_slice(&self.message.data);
+                data
+            }
+            SignDigestType::Eip712 => {
+                let value = serde_json::from_slice(&self.message.data).unwrap_or_default();
+                if let Ok(message) = parse_eip712_json(&value) {
+                    return message.hash();
+                }
+                Vec::new()
+            }
+            SignDigestType::Base58 => {
+                // Check if the data is a valid base58 string in utf8
+                if let Ok(string_data) = String::from_utf8(self.message.data.clone()) {
+                    if bs58::decode(string_data.as_bytes()).into_vec().is_ok() {
+                        return self.message.data.clone();
+                    }
+                }
+                Vec::new()
+            }
+        }
+    }
+
     pub fn get_result(&self, data: &[u8]) -> String {
         match self.message.sign_type {
             SignDigestType::Sign | SignDigestType::Eip191 | SignDigestType::Eip712 => hex::encode_prefixed(data),
@@ -65,7 +91,6 @@ mod tests {
     use super::*;
     use crate::message::sign_type::SignDigestType;
     use alloy_primitives::hex;
-    use gem_evm::eip712::EIP712TypedValue;
 
     #[test]
     fn test_eip191() {
@@ -126,84 +151,5 @@ mod tests {
         let result = decoder.get_result(result_data);
 
         assert_eq!(result, "3LRFsmWKLfsR7G5PqjytR");
-    }
-
-    #[test]
-    fn test_eip712() {
-        // Load the JSON content from the file
-        let json_data_str = include_str!("test/uniswap_permit2.json");
-        let data = json_data_str.as_bytes().to_vec();
-
-        let decoder = SignMessageDecoder::new(SignMessage {
-            sign_type: SignDigestType::Eip712,
-            data: data.clone(),
-        });
-
-        match decoder.preview() {
-            Ok(MessagePreview::EIP712(message)) => {
-                assert_eq!(message.domain.name, "Permit2");
-                assert_eq!(message.domain.chain_id, 1);
-                assert_eq!(message.domain.verifying_contract.to_lowercase(), "0x000000000022d473030f116ddee9f6b43ac78ba3");
-                assert_eq!(message.message.len(), 3);
-                assert_eq!(message.message[0].name, "details");
-
-                match &message.message[0].value {
-                    EIP712TypedValue::Struct { fields } => {
-                        assert_eq!(fields.len(), 4); // token, amount, expiration, nonce
-
-                        // 1.1 token (address)
-                        assert_eq!(fields[0].name, "token");
-                        match &fields[0].value {
-                            EIP712TypedValue::Address { value } => assert_eq!(value, "0xdAC17F958D2ee523a2206206994597C13D831ec7"),
-                            _ => panic!("Incorrect type for details.token"),
-                        }
-                        // 1.2 amount (uint160 - parsed as Uint256 for now)
-                        // We parse uint160 as Uint256 { value: String } because the JSON value is a string.
-                        assert_eq!(fields[1].name, "amount");
-                        match &fields[1].value {
-                            EIP712TypedValue::Uint256 { value } => assert_eq!(value, "1461501637330902918203684832716283019655932542975"),
-                            _ => panic!("Incorrect type for details.amount"),
-                        }
-                        // 1.3 expiration (uint48 - parsed as Uint256 for now)
-                        assert_eq!(fields[2].name, "expiration");
-                        match &fields[2].value {
-                            EIP712TypedValue::Uint256 { value } => assert_eq!(value, "1732780554"),
-                            _ => panic!("Incorrect type for details.expiration"),
-                        }
-                        // 1.4 nonce (uint48 - parsed as Uint256 for now)
-                        assert_eq!(fields[3].name, "nonce");
-                        match &fields[3].value {
-                            EIP712TypedValue::Uint256 { value } => assert_eq!(value, "0"),
-                            _ => panic!("Incorrect type for details.nonce"),
-                        }
-                    }
-                    _ => panic!("Incorrect type for details field"),
-                }
-
-                assert_eq!(message.message[1].name, "spender");
-                match &message.message[1].value {
-                    EIP712TypedValue::Address { value } => {
-                        assert_eq!(value.to_lowercase(), "0x3fc91a3afd70395cd496c647d5a6cc9d4b2b7fad");
-                    }
-                    _ => panic!("Expected spender field to be an Address"),
-                }
-
-                assert_eq!(message.message[2].name, "sigDeadline");
-                match &message.message[2].value {
-                    EIP712TypedValue::Uint256 { value } => {
-                        assert_eq!(value, "1730190354");
-                    }
-                    _ => panic!("Expected sigDeadline field to be a Uint256"),
-                }
-            }
-            Ok(_) => panic!("Expected EIP712 preview, got Text"),
-            Err(e) => panic!("Preview failed for EIP712: {:?}", e),
-        }
-
-        // Test get_result: Should just hex encode the input data for EIP712
-        let result_data = b"some_bytes_to_encode";
-        let result = decoder.get_result(result_data);
-
-        assert_eq!(result, hex::encode_prefixed(result_data));
     }
 }
